requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
pydantic-settings>=2.2.1
python-multipart>=0.0.9
slowapi>=0.1.9

sentence-transformers>=3.0.1
torch>=2.2.0
faiss-cpu>=1.8.0
numpy>=1.26.0

pypdf>=4.2.0
tiktoken>=0.7.0

openai>=1.40.0
httpx>=0.27.0
python-dotenv>=1.0.1
google-generativeai>=0.8.3

orjson>=3.10.0
cachetools>=5.3.3
streamlit>=1.37.0
requests>=2.32.0
